Updated physics.h and cpp with structs and coordinate transformation functions

+++ src/Constants.h
@@ -1,2 +1,2 @@
 #pragma once
-const float dt = 0.01f;
+const double dt = 0.01f;
